WIP: Add some status types
For all status types, we should define an explicit data format. The
purpose of this format is to allow different front ends to render the
output. The status types currently defined live in
riak_core_status_types.hrl.

riak_core_console implements a formatting callback, write_status/2, for
the console. So far it is not completely implemented. I only implemented
enough of the callback to handle tags used in
riak_core_status:ring_ready/0. Similarly an html output could be
generated by implementing a write_status/2 callback that output html.
The generic part of the parser exists in riak_core_status:parse/3. It is
called with a specific formatting implementation as well as a context
specific to the formatter, for accumulation during the parsing process.