Add autosizing to riak_core_console_table
When calling riak_core_console_table:print/2 without a spec including
sizes, the table elements will be sized according to the maximum length
of element in each row. If that all fits within the console width,
including borders, then it is printed. If it does not fit, the elements
are reduced in size until everything fits on screen. If the console
width is not detected with io:columns() then the default value,
?MAX_LINE_LEN=100 is used instead.